HIVE-4848 : std, stddev and stddev_pop aggregates on double/float fail to vectorize (Remus Rusanu via Ashutosh Chauhan)
git-svn-id: https://svn.apache.org/repos/asf/hive/branches/vectorization@1504216 13f79535-47bb-0310-9956-ffa450edef68

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/vector/VectorizationContext.java
@@ -1091,9 +1091,9 @@ public class VectorizationContext {
     {"std",       "Long",   VectorUDAFStdPopLong.class},
     {"stddev",    "Long",   VectorUDAFStdPopLong.class},
     {"stddev_pop","Long",   VectorUDAFStdPopLong.class},
-    {"std",       "Long",   VectorUDAFStdPopDouble.class},
-    {"stddev",    "Long",   VectorUDAFStdPopDouble.class},
-    {"stddev_pop","Long",   VectorUDAFStdPopDouble.class},
+    {"std",       "Double", VectorUDAFStdPopDouble.class},
+    {"stddev",    "Double", VectorUDAFStdPopDouble.class},
+    {"stddev_pop","Double", VectorUDAFStdPopDouble.class},
     {"stddev_samp","Long",  VectorUDAFStdSampLong.class},
     {"stddev_samp","Double",VectorUDAFStdSampDouble.class},
   };
